+++ milvus/client/grpc_client/interceptor/__init__.py
@@ -1,0 +1,4 @@
+from . import header_client_interceptor
+
+
+__all__ = ['header_client_interceptor']