+++ milvus/client/types.py
@@ -9,30 +9,6 @@ class Status:
     """
 
     SUCCESS = 0
-    UNEXPECTED_ERROR = 1
-    CONNECT_FAILED = 2
-    PERMISSION_DENIED = 3
-    COLLECTION_NOT_EXISTS = 4
-    ILLEGAL_ARGUMENT = 5
-    ILLEGAL_RANGE = 6
-    ILLEGAL_DIMENSION = 7
-    ILLEGAL_INDEX_TYPE = 8
-    ILLEGAL_COLLECTION_NAME = 9
-    ILLEGAL_TOPK = 10
-    ILLEGAL_ROWRECORD = 11
-    ILLEGAL_VECTOR_ID = 12
-    ILLEGAL_SEARCH_RESULT = 13
-    FILE_NOT_FOUND = 14
-    META_FAILED = 15
-    CACHE_FAILED = 16
-    CANNOT_CREATE_FOLDER = 17
-    CANNOT_CREATE_FILE = 18
-    CANNOT_DELETE_FOLDER = 19
-    CANNOT_DELETE_FILE = 20
-    BUILD_INDEX_ERROR = 21
-    ILLEGAL_NLIST = 22
-    ILLEGAL_METRIC_TYPE = 23
-    OUT_OF_MEMORY = 24
 
     def __init__(self, code=SUCCESS, message="Success"):
         self.code = code
@@ -77,12 +53,3 @@ class DataType(IntEnum):
     # VECTOR = 200
 
     UNKNOWN = 999
-
-
-class RangeType(IntEnum):
-    LT = 0   # less than
-    LTE = 1  # less than or equal
-    EQ = 2   # equal
-    GT = 3   # greater than
-    GTE = 4  # greater than or equal
-    NE = 5   # not equal